+++ packages/adapter-neo4j/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@next-auth/neo4j-adapter",
-  "version": "1.0.6",
-  "description": "neo4j adapter for next-auth.",
+  "name": "@auth/neo4j-adapter",
+  "version": "0.0.0",
+  "description": "neo4j adapter for Auth.js",
   "homepage": "https://authjs.dev",
   "repository": "https://github.com/nextauthjs/next-auth",
   "bugs": {
@@ -11,7 +11,19 @@
   "contributors": [
     "Balázs Orbán <info@balazsorban.com>"
   ],
-  "main": "dist/index.js",
+  "type": "module",
+  "types": "./index.d.ts",
+  "files": [
+    "*.js",
+    "*.d.ts*",
+    "src"
+  ],
+  "exports": {
+    ".": {
+      "types": "./index.d.ts",
+      "import": "./index.js"
+    }
+  },
   "license": "ISC",
   "keywords": [
     "next-auth",
@@ -28,24 +40,18 @@
     "test": "./tests/test.sh",
     "build": "tsc"
   },
-  "files": [
-    "README.md",
-    "dist"
-  ],
   "peerDependencies": {
-    "neo4j-driver": "^4.0.0 || ^5.7.0",
-    "next-auth": "^4"
+    "neo4j-driver": "^4.0.0 || ^5.7.0"
   },
   "devDependencies": {
     "@next-auth/adapter-test": "workspace:*",
     "@next-auth/tsconfig": "workspace:*",
     "@types/uuid": "^8.3.3",
     "jest": "^27.4.3",
-    "neo4j-driver": "^5.7.0",
-    "next-auth": "workspace:*"
+    "neo4j-driver": "^5.7.0"
   },
   "dependencies": {
-    "uuid": "^8.3.2"
+    "@auth/core": "workspace:*"
   },
   "jest": {
     "preset": "@next-auth/adapter-test/jest"
