+++ packages/adapter-neo4j/src/index.ts
@@ -9,17 +9,13 @@
  * ## Installation
  *
  * ```bash npm2yarn2pnpm
- * npm install next-auth @next-auth/neo4j-adapter neo4j-driver
+ * npm install @auth/neo4j-adapter neo4j-driver
  * ```
  *
- * @module @next-auth/neo4j-adapter
+ * @module @auth/neo4j-adapter
  */
-import type { Session } from "neo4j-driver"
-import type { Adapter } from "next-auth/adapters"
-import { v4 as uuid } from "uuid"
-
-import { client, format } from "./utils"
-export { format }
+import { type Session, isInt, integer } from "neo4j-driver"
+import type { Adapter } from "@auth/core/adapters"
 
 /** This is the interface of the Neo4j adapter options. The Neo4j adapter takes a {@link https://neo4j.com/docs/bolt/current/driver-api/#driver-session Neo4j session} as its only argument. */
 export interface Neo4jOptions extends Session {}
@@ -31,7 +27,7 @@ export interface Neo4jOptions extends Session {}
  *
  * ```javascript title="pages/api/auth/[...nextauth].js"
  * import neo4j from "neo4j-driver"
- * import { Neo4jAdapter } from "@next-auth/neo4j-adapter"
+ * import { Neo4jAdapter } from "@auth/neo4j-adapter"
  *
  * const driver = neo4j.driver(
  *   "bolt://localhost",
@@ -134,7 +130,7 @@ export function Neo4jAdapter(session: Session): Adapter {
 
   return {
     async createUser(data) {
-      const user: any = { id: uuid(), ...data }
+      const user = { id: crypto.randomUUID(), ...data }
       await write(`CREATE (u:User $data)`, user)
       return user
     },
@@ -285,6 +281,72 @@ export function Neo4jAdapter(session: Session): Adapter {
         data
       )
       return format.from<any>(result?.properties)
+    },
+  }
+}
+
+// https://github.com/honeinc/is-iso-date/blob/master/index.js
+const isoDateRE =
+  /(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))/
+
+function isDate(value: any) {
+  return value && isoDateRE.test(value) && !isNaN(Date.parse(value))
+}
+
+export const format = {
+  /** Takes a plain old JavaScript object and turns it into a Neo4j compatible object */
+  to(object: Record<string, any>) {
+    const newObject: Record<string, unknown> = {}
+    for (const key in object) {
+      const value = object[key]
+      if (value instanceof Date) newObject[key] = value.toISOString()
+      else newObject[key] = value
+    }
+    return newObject
+  },
+  /** Takes a Neo4j object and returns a plain old JavaScript object */
+  from<T = Record<string, unknown>>(object?: Record<string, any>): T | null {
+    const newObject: Record<string, unknown> = {}
+    if (!object) return null
+    for (const key in object) {
+      const value = object[key]
+      if (isDate(value)) {
+        newObject[key] = new Date(value)
+      } else if (isInt(value)) {
+        if (integer.inSafeRange(value)) newObject[key] = value.toNumber()
+        else newObject[key] = value.toString()
+      } else {
+        newObject[key] = value
+      }
+    }
+
+    return newObject as T
+  },
+}
+
+function client(session: Session) {
+  return {
+    /** Reads values from the database */
+    async read<T>(statement: string, values?: any): Promise<T | null> {
+      const result = await session.readTransaction((tx) =>
+        tx.run(statement, values)
+      )
+
+      return format.from<T>(result?.records[0]?.get(0)) ?? null
+    },
+    /**
+     * Reads/writes values from/to the database.
+     * Properties are available under `$data`
+     */
+    async write<T extends Record<string, any>>(
+      statement: string,
+      values: T
+    ): Promise<any> {
+      const result = await session.writeTransaction((tx) =>
+        tx.run(statement, { data: format.to(values) })
+      )
+
+      return format.from<T>(result?.records[0]?.toObject())
     },
   }
 }
