+++ turbo.json
@@ -61,7 +61,7 @@
         "@auth/firebase-adapter#build",
         "@auth/mikro-orm-adapter#build",
         "@auth/mongodb-adapter#build",
-        "@next-auth/neo4j-adapter#build",
+        "@auth/neo4j-adapter#build",
         "@next-auth/pouchdb-adapter#build",
         "@next-auth/sequelize-adapter#build",
         "@next-auth/supabase-adapter#build",
@@ -84,7 +84,7 @@
         "@auth/firebase-adapter#build",
         "@auth/mikro-orm-adapter#build",
         "@auth/mongodb-adapter#build",
-        "@next-auth/neo4j-adapter#build",
+        "@auth/neo4j-adapter#build",
         "@next-auth/pouchdb-adapter#build",
         "@next-auth/sequelize-adapter#build",
         "@next-auth/supabase-adapter#build",
